Add AIMessageType to scripting system - Add AIMessageTypeLiteral to ScriptCompilerCommon.h - Allow tokenizing, parsing, and generation of C++ output for new AIMessageType in the script compiler - Add interface for creating CRC hash using an std::string in FileSystem.cpp/.h

+++ Engine/Source/Kargono/AI/AIService.cpp
@@ -97,7 +97,7 @@ namespace Kargono::AI
 		
 		
 	}
-	void AIService::SendAIMessage(uint64_t messageType, UUID senderEntity, UUID receiverEntity, float messageDelay)
+	void AIService::SendAIMessage(uint32_t messageType, UUID senderEntity, UUID receiverEntity, float messageDelay)
 	{
 		// Route message to either the HandleAIMessage() function or the HandleDelayedMessages() function
 		//		based on the messageDelay
